ai_agent_tutorials/ai_3dpygame_r1/email_to_ERP_Agent/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.9
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
email-validator==2.0.0
requests==2.29.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
langchain==0.0.335
openai==1.2.4
tiktoken==0.5.1
jinja2==3.1.2
aiosmtplib==2.0.2
tenacity==8.2.3
